refactor: move product domain to product package
