+++ app/src/main/java/com/codesoom/assignment/global/error/ControllerErrorAdvice.java
@@ -1,6 +1,6 @@
 package com.codesoom.assignment.global.error;
 
-import com.codesoom.assignment.errors.ProductNotFoundException;
+import com.codesoom.assignment.product.application.ProductNotFoundException;
 import com.codesoom.assignment.user.application.UserEmailDuplicationException;
 import com.codesoom.assignment.user.application.UserNotFoundException;
 import org.springframework.http.HttpStatus;
