+++ app/src/main/java/com/codesoom/assignment/product/application/ProductService.java
@@ -1,9 +1,9 @@
-package com.codesoom.assignment.application;
+package com.codesoom.assignment.product.application;
 
-import com.codesoom.assignment.domain.Product;
-import com.codesoom.assignment.domain.ProductRepository;
-import com.codesoom.assignment.dto.ProductData;
-import com.codesoom.assignment.errors.ProductNotFoundException;
+import com.codesoom.assignment.product.domain.Product;
+import com.codesoom.assignment.product.domain.ProductRepository;
+import com.codesoom.assignment.product.dto.ProductData;
+import com.codesoom.assignment.product.application.ProductNotFoundException;
 import com.github.dozermapper.core.Mapper;
 import org.springframework.stereotype.Service;
