+++ app/src/test/java/com/codesoom/assignment/product/controller/ProductControllerTest.java
@@ -1,9 +1,10 @@
-package com.codesoom.assignment.controllers;
+package com.codesoom.assignment.product.controller;
 
-import com.codesoom.assignment.application.ProductService;
-import com.codesoom.assignment.domain.Product;
-import com.codesoom.assignment.dto.ProductData;
-import com.codesoom.assignment.errors.ProductNotFoundException;
+import com.codesoom.assignment.product.application.ProductService;
+import com.codesoom.assignment.product.controller.ProductController;
+import com.codesoom.assignment.product.domain.Product;
+import com.codesoom.assignment.product.dto.ProductData;
+import com.codesoom.assignment.product.application.ProductNotFoundException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
