+++ app/src/test/java/com/codesoom/assignment/product/domain/ProductTest.java
@@ -1,5 +1,6 @@
-package com.codesoom.assignment.domain;
+package com.codesoom.assignment.product.domain;
 
+import com.codesoom.assignment.product.domain.Product;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
